Thierry

+++ Controllers/Controller_user.php
@@ -12,13 +12,29 @@ class Controller_user extends Controller
     }
 
 
+    // public function action_all_users()
+    // {
+    //     $m=User::get_model();
+    //     $data=['users'=>$m->get_all_users()];
+    //     $this->render("all_users",$data);
+    // }
     public function action_all_users()
     {
         $m=User::get_model();
-        $data=['users'=>$m->get_all_users()];
-        $this->render("all_users",$data);
+        $users = $m->get_all_users();
+        foreach ($users as $user) {
+            $lastActivityTimestamp = strtotime($user->lastActivityTime);
+            $currentTimestamp = time();
+            $timeDifference = $currentTimestamp - $lastActivityTimestamp;
+    
+            $user->active = ($timeDifference <= 300); // 5 minutes en secondes (5 * 60 = 300)
+        }
 
+        $data=['users'=> $users];
+        $this->render("all_users",$data);
     }
+
+
     public function action_user_profile()
     {
         $m=User::get_model();
@@ -45,31 +61,95 @@ class Controller_user extends Controller
     public function action_public_profile()
     {
         $m=User::get_model();
-        $data=['user'=>$m->get_public_profile(),
-               'followers'=>$m->get_followers_number_public(),
+        $user = $m->get_public_profile();
+        
+        $lastActivityTimestamp = strtotime($user['user_info'][0]->lastActivityTime);
+        $currentTimestamp = time();
+        $timeDifference = $currentTimestamp - $lastActivityTimestamp;
+
+        $user['user_info'][0]->active = ($timeDifference <= 300); // 5 minutes en secondes (5 * 60 = 300)
+
+
+        $data=[
+            // 'user'=>$m->get_public_profile(),
+                'user'=>$user,
+                'followers'=>$m->get_followers_number_public(),
                'followed'=>$m->get_followed_number_public(),
                'isFollowing'=>$m->get_is_following()];
-        $this->render("public_profile", $data);
+        $this->
+        render("public_profile", $data);
     }
+    
+    // public function action_leaderboard()
+    // {
+    //     $m=User::get_model();
+    //     $data=['users'=>$m->get_leaderboard()];
+    //     $this->render("leaderboard", $data);
+    // }
+
     public function action_leaderboard()
-    {
-        $m=User::get_model();
-        $data=['users'=>$m->get_leaderboard()];
-        $this->render("leaderboard", $data);
+{
+    $m = User::get_model();
+    $users = $m->get_leaderboard();
+
+    foreach ($users as $user) {
+        $lastActivityTimestamp = strtotime($user->lastActivityTime);
+        $currentTimestamp = time();
+        $timeDifference = $currentTimestamp - $lastActivityTimestamp;
+
+        $user->active = ($timeDifference <= 300); // 5 minutes en secondes (5 * 60 = 300)
     }
 
+    $data = ['users' => $users];
+    $this->render("leaderboard", $data);
+}
+
+    // public function action_all_followers()
+    // {
+    //     $m=User::get_model();
+    //     $data=['follow'=>$m->get_all_followers(),
+    //            'message'=>'Liste des abonnés'];
+    //     $this->render("all_follow", $data);
+    // }
     public function action_all_followers()
     {
         $m=User::get_model();
-        $data=['follow'=>$m->get_all_followers(),
+        $followers = $m->get_all_followers();
+
+        foreach ($followers as $follow) {
+            $lastActivityTimestamp = strtotime($follow->lastActivityTime);
+            $currentTimestamp = time();
+            $timeDifference = $currentTimestamp - $lastActivityTimestamp;
+    
+            $follow->active = ($timeDifference <= 300); // 5 minutes en secondes (5 * 60 = 300)
+        }
+
+        $data=['follow'=>$followers,
                'message'=>'Liste des abonnés'];
         $this->render("all_follow", $data);
     }
 
+    // public function action_all_followed()
+    // {
+    //     $m=User::get_model();
+    //     $data=['follow'=>$m->get_all_followed(),
+    //            'message'=>'Liste des abonnments'];
+    //     $this->render("all_follow", $data);
+    // }
     public function action_all_followed()
     {
         $m=User::get_model();
-        $data=['follow'=>$m->get_all_followed(),
+        $followed = $m->get_all_followed();
+
+        foreach ($followed as $follow) {
+        $lastActivityTimestamp = strtotime($follow->lastActivityTime);
+        $currentTimestamp = time();
+        $timeDifference = $currentTimestamp - $lastActivityTimestamp;
+
+        $follow->active = ($timeDifference <= 300); // 5 minutes en secondes (5 * 60 = 300)
+        }
+
+        $data=['follow'=>$followed,
                'message'=>'Liste des abonnments'];
         $this->render("all_follow", $data);
     }
@@ -121,15 +201,40 @@ class Controller_user extends Controller
                     document.location.href = '?controller=user&action=user_profile'
                 </script>
                 ";
-            } elseif ($imageSize > 1200000){
-                echo 
-                "
-                <script>
-                    alert('Image trop lourde ! 1,2 Mo max !')
-                    document.location.href = '?controller=user&action=user_profile'
-                </script>
-                ";
+
+                /* ------------------------------- ancien code ------------------------------ */
+            // } elseif ($imageSize > 1200000){
+            //     echo 
+            //     "
+            //     <script>
+            //         alert('Image trop lourde ! 1,2 Mo max !')
+            //         document.location.href = '?controller=user&action=user_profile'
+            //     </script>
+            //     ";
+            // } else {
+                /* ------------------------------- ancien code ------------------------------ */
+
             } else {
+                $maxFileSize = 1200000; // 1,2 Mo
+    
+                if ($imageSize > $maxFileSize) {
+                    $image = new Imagick($tmpName);
+
+                    if($this->isMobileImage($image)) {
+                        $image->trimImage(0);
+                    }
+    
+                    $compressionQuality = 80; // Qualité de compression (entre 0 et 100)
+                    $image->setImageCompressionQuality($compressionQuality);
+    
+                    $newImageName = $username."_".date('Y.m.d')."_".date('h.i.sa').".".$imageExtension;
+                    $image->writeImage('Public/img/' . $newImageName);
+    
+                    $image->destroy();
+                } else {
+                    $newImageName = $username."_".date('Y.m.d')."_".date('h.i.sa').".".$imageExtension;
+                    move_uploaded_file($tmpName, 'Public/img/' . $newImageName);
+                }
 
                 $m=User::get_model();
                 $oldImageName = $m->get_profile_picture($user_id);
@@ -138,7 +243,7 @@ class Controller_user extends Controller
                 
 
                 // Delete old image if exists
-                if($oldImageName !== null && file_exists('Public/img/' . $oldImageName)) {
+                if($oldImageName !== null && $oldImageName !== 'noprofile.png' && file_exists('Public/img/' . $oldImageName)) {
                     unlink('Public/img/' . $oldImageName);
                 }
 
@@ -146,6 +251,7 @@ class Controller_user extends Controller
                 $newImageName.=".".$imageExtension;
                 $m=User::get_model();
                 $m->set_profile_picture($newImageName);
+                $_SESSION['image_name'] = $newImageName;
                 move_uploaded_file($tmpName, 'Public/img/' . $newImageName);
                 echo 
                 "
@@ -159,6 +265,22 @@ class Controller_user extends Controller
         }
     }
 
-    
+    private function isMobileImage($image) {
+        $width = $image->getImageWidth();
+        $height = $image->getImageHeight();
+
+        // Vous pouvez ajuster ces valeurs en fonction de votre expérience
+        $mobileWidthThreshold = 1000;
+        $mobileHeightThreshold = 1000;
+
+        if($width <= $mobileWidthThreshold && $height <= $mobileHeightThreshold) {
+            $size = min($width, $height);
+            $image->cropImage($size, $size, 0, 0);
+            return true;
+        }
+
+        return false;
+    }
+
 
 }
